:recycle: Rename to weekly calendar state
