+++ lib/domain/calendar/weekly_calendar_state.dart
